requirements.txt
python-dotenv==1.2.1
streamlit==1.53.1
openai==2.16.0